+++ src/Resources/contao/languages/de/tl_backendmenue_bereiche.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 $GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['name'] = ['Name', 'Name des Menübereichs'];
-$GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['icon'] = ['Icon', 'Icon für den Menübereich (Font Awesome 6 oder Contao-Icon)'];
+$GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['icon'] = ['Icon', 'Icon für den Menübereich (Font Awesome oder Contao-Backend-Icon); wird vor dem Bereichsnamen in der Navigation angezeigt'];
 $GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['position'] = ['Position', 'Sortierposition des Bereichs'];
 $GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['edit'] = ['Module verwalten', 'Die Modul-Zuordnungen des Bereichs ID %s verwalten'];
 $GLOBALS['TL_LANG']['tl_backendmenue_bereiche']['editheader'] = ['Bereich bearbeiten', 'Die Einstellungen des Bereichs ID %s bearbeiten'];